Testing out different values for animal types
using single instance for Enum.AnimalType value constantly being checked inside collider logic of run state and walk state instantiated at on enter state

Horse is now a medium animal so as to let its run animation be applied.
