+++ CarmaRoad/Assets/Animations/Animals/BlackCat/Idle.anim.meta
