+++ CarmaRoad/Assets/Animations/Animals/BlackCat/Run.anim.meta
