+++ CarmaRoad/Assets/Prefabs/Hyena.prefab
@@ -152,7 +152,7 @@ CircleCollider2D:
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
-  m_Radius: 5
+  m_Radius: 4
 --- !u!95 &1256627704242155793
 Animator:
   serializedVersion: 5
