+++ CarmaRoad/Assets/Scripts/Misc/ScriptableObject/BaseAnimal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e50b55746a52f642b238c5549a9616a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
