+++ CarmaRoad/Assets/Scripts/Misc/ScriptableObject/BaseVehicle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93d4bd8693079bb4a8407a3b6f2026d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
