+++ CarmaRoad/Assets/Scripts/Services/AnimalService/CreateAnimalService.cs
@@ -7,7 +7,7 @@ namespace CarmaRoad
     {
         private AnimalController animalController;
 
-        public AnimalController CreateAnimal(BaseAnimalObject baseAnimal, Vector2 spawnPoint, Enum.AnimalSpawnPosition animalSpawnPosition)
+        public AnimalController CreateAnimal(BaseAnimal baseAnimal, Vector2 spawnPoint, Enum.AnimalSpawnPosition animalSpawnPosition)
         {
             AnimalModel animalModel = new AnimalModel(baseAnimal);
             animalController = new AnimalController(animalModel, baseAnimal.animalPrefab, spawnPoint, animalSpawnPosition);
